b1_favorite_ms/favoritePlaces/serializers.py
from .models import Favourite
from rest_framework import serializers


class FavouriteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Favourite
        fields = ('id', 'user_id', 'place_id', 'comment')
        