requirements.txt
Django>=1.8<2.0
djangorestframework==3.7.7
psycopg2